atualizando

+++ base/views.py
@@ -1,10 +1,12 @@
 from django.shortcuts import render, redirect
 from django.db.models import Q
+from django.http import HttpResponse
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
-from django.contrib import messages
+from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth.decorators import login_required
-from .models import Room, Topic
+from django.contrib import messages
+from .models import Room, Topic, Message
 from .forms import RoomForm
 
 # Create your views here.
@@ -15,10 +17,15 @@ from .forms import RoomForm
     #{'id':3, 'name':'aprender css'},
 #]
 
+
 def loginPagina(request):
 
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
         try:
             user = User.objects.get(username=username)
@@ -31,7 +38,7 @@ def loginPagina(request):
             login(request, user)
             return redirect('home')
 
-    context = {}
+    context = {'page':page}
     return render(request, 'base/login_register.html', context)
 
 
@@ -40,6 +47,24 @@ def logoutUser(request):
     return redirect('home')
 
 
+def registerPagina(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Algo de errado ocorreu ao se registrar.')
+
+    return render(request, 'base/login_register.html', {'form':form})
+
+
+@login_required(login_url='login-pagina')
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
 
@@ -49,18 +74,42 @@ def home(request):
         Q(discription__icontains=q)
         )
 
+    user_rooms = Room.objects.filter(host=request.user)
+    num_rooms = user_rooms.count()
+
     topics = Topic.objects.all()
     room_count = rooms.count()
+    room_messages = Message.objects.filter(Q(room__topic__name__icontains=q))
 
-    context = {'rooms':rooms, 'topics':topics, 'room_count':room_count}
+    context = {'rooms':rooms, 'topics':topics, 'room_count':room_count, 'num_rooms':num_rooms, 'room_messages':room_messages}
     return render(request, 'base/home.html', context)
 
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
-    context = {'room':room}
+    room_messages = room.message_set.all().order_by('-created')
+    participants = room.participants.all
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user=request.user,
+            room=room,
+            body=request.POST.get('body')
+        )
+        room.participants.add(request.user)
+        return redirect('room', pk=room.id)
+        
+    context = {'room':room, 'room_messages':room_messages, 'participants':participants}
 
     return render(request, 'base/room.html', context)
+
+
+def userProfile(request, pk):
+    user = User.objects.get(id=pk)
+    rooms = user.room_set.all()
+    room_messages = user.message_set.all()
+    topics = Topic.objects.all()
+    context = {'user':user, 'rooms':rooms, 'room_messages':room_messages, 'topics':topics}
+    return render(request, 'base/profile.html', context)
 
 
 @login_required(login_url='login-pagina')
@@ -81,6 +130,9 @@ def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
 
+    if request.user != room.host:
+        return HttpResponse('Este tópico é domínio de outra conta')
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -92,7 +144,24 @@ def updateRoom(request, pk):
 @login_required(login_url='login-pagina')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('Este tópico é domínio de outra conta')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render(request, 'base/delete.html', {'obj':room})
+
+
+@login_required(login_url='login-pagina')
+def deleteMessage(request, pk):
+    message = Message.objects.get(id=pk)
+
+    if request.user != message.user:
+        return HttpResponse('Este mensagem é domínio de outra conta')
+
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+    return render(request, 'base/delete.html', {'obj':message})
